+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] ; 2025-06-25
+
+### Added
+
+- Suite of 4 core validation tests
+- Validation test orchestrator
+
 ## [0.2.1] ; 2025-06-25
 
 ### Changed